# Conflicts:
#	src/main/java/iga/TransfusionSanguine/Repositories/PersonneRepository.java

+++ src/main/java/iga/TransfusionSanguine/Repositories/PersonnelRepository.java
@@ -1,5 +1,6 @@
 package iga.TransfusionSanguine.Repositories;
 
+import iga.TransfusionSanguine.Entities.Personne;
 import iga.TransfusionSanguine.Entities.Personnel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -7,5 +8,7 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 
 @Repository
 @CrossOrigin("*")
-public interface PersonnelRepository extends JpaRepository<Personnel,Integer> {
+public interface PersonnelRepository extends JpaRepository<Personnel,Long>
+{
+
 }
